backend/requirements.txt
pymongo==4.3.3
fastapi==0.95.0
uvicorn==0.21.1
motor==3.1.2
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
python-dateutil==2.8.2
autopep8==2.0.2
isort==5.12.0

pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
pytest-cov==4.0.0
pytest-watch==4.2.0
pytest-lazy-fixture==0.6.3
pytest-random-order==1.1.0
pytest-xdist==3.3.0
httpx==0.24.0